debug

+++ lora_attack/eval/eval.py
@@ -63,7 +63,7 @@ if __name__ == '__main__':
     logger.info(json.dumps(args, indent=4))
     model_name = args['model_dir']['name']
     model = AutoModelForCausalLM.from_pretrained(model_name, device_map='cuda:0',
-                                                 attn_implementation="flash_attention_2", torch_dtype=torch.float16,
+                                                 attn_implementation="flash_attention_2", torch_dtype=torch.bfloat16,
                                                  token=hf_access_token)
     tokenizer = AutoTokenizer.from_pretrained(model_name, token=hf_access_token, padding_side="left")
     tokenizer.pad_token = tokenizer.eos_token
@@ -82,7 +82,7 @@ if __name__ == '__main__':
             #    )
             model = PeftModel.from_pretrained(model=model, model_id=args['adapter_dir'],
                                               device_map='cuda:0', attn_implementation="flash_attention_2",
-                                              torch_dtype=torch.float16,
+                                              torch_dtype=torch.bfloat16,
                                               token=hf_access_token,
                                               adapter_name="task",
                                               quantization_config=quantization_config)
